requirements.txt
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
jmespath==1.0.0
aws-psycopg2
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
